requirements.txt
django==3.2.6
django-crispy-forms==1.12.0